Refactor FetchJokeTask as external class for testability
