+++ app/src/androidTest/java/com/udacity/gradle/builditbigger/FetchJokeTest.java
@@ -1,0 +1,10 @@
+package com.udacity.gradle.builditbigger;
+
+import android.test.AndroidTestCase;
+
+/**
+ * Created by geraud on 03/03/16.
+ */
+public class FetchJokeTest extends AndroidTestCase {
+
+}
